combining account settings and edit into one file
