+++ client/src/App.tsx
@@ -3,8 +3,7 @@ import { Login } from './pages/Login/Login';
 import { Homepage } from './pages/Home/homepage';
 import { Welcome } from './pages/Home/Welcome';
 import SoilGame from './pages/SoilGame/index';
-import { AccountSettings } from './pages/AccountSettings/AccountSettings';
-import { AccountSettingsEdit } from './pages/AccountSettings/AccountSettingsEdit';
+import { AccountSettings, AccountSettingsEdit } from './pages/AccountSettings/AccountSettings';
 import { WaterGame } from './pages/WaterGame/WaterGame';
 import './App.css';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
